requirements_vertex_ai.txt
# Vertex AI Fine-Tuning Requirements
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1
vertexai>=1.38.0

# Additional dependencies
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0 